Added animation when computer makes a move
Previously, when it was the computer's turn to move and the
ChooseMove() method was called, the GUI just made the hole
in the GameBoard change color and play immediately returned
to the human player. Now a yellow AICirlce appears at the top of
the board and slowly moves to the chosen column before the
selected hole color changes.

Other notable change:

- pickColumn method within BoardModel no longer calls switchTurn().
  switchTurn() must be invoked separately. This allows the AICircle
  to display until while it is the computer's turn, then revert
  back to the normal (player's) Circle.