Bug fixes.
Fixed 2 bugs to make this a stable version of ConnectFour.

- Moved "findWinner()" block to the end of the animation
after the computer makes a play. Otherwise, the token will appear
in its new hole while the animation of it moving into position is
still running.

- Corrected the placement of switcTurn() inside the AIPlayer's
algorithm. Now switchTurn is only executed if a valid column has
been chosen. This bug caused lots of strange behavior and runtime
errors in the previous version.